classe TelloZune melhorada
